+++ CHANGELOG.md
@@ -4,6 +4,19 @@ This changelog follows the patterns described here: <https://keepachangelog.com/
 
 Subheadings to categorize changes are `added, changed, deprecated, removed, fixed, security`.
 
+## 0.8.10
+
+### added
+
+- `.add_sendonly_protocol()` for protocols that are never read.
+- `.add_readonly_bounded_protocol(usize)` for protocols that are never sent, with a bounded receiver.
+- `.add_readonly_unbounded_protocol()` for protocols that are never sent, with an unbounded receiver.
+
+### fixed
+
+- README compatibility table
+
+
 ## 0.8.9
 
 ### fixed